@@ -2,6 +2,7 @@
   "name": "mern-auth",
   "version": "1.0.0",
   "description": "Login/Registration authentication App",
+  "type": "module",
   "main": "server.js",
   "scripts": {
     "client-install": "npm install --prefix client",
@@ -15,14 +16,14 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.19.0",
+    "axios": "^1.6.4",
+    "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "classnames": "^2.2.6",
     "concurrently": "^5.0.2",
     "dotenv": "^16.3.1",
     "express": "^4.17.1",
-    "is-empty": "^1.2.0",
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^2.2.0",
     "mongoose": "^5.8.2",
@@ -42,6 +43,6 @@
     "nodemon": "^2.0.2"
   },
   "engines": {
-    "node": "12.13.1"
+    "node": "18.17.1"
   }
 }
