@@ -1,4 +1,4 @@
-module.exports = {
-    mongoURI:  "mongodb://localhost:27017/mern-auth",
-    secretOrKey: "secret"
-  };
+export const key = {
+  mongoURI: "mongodb://localhost:27017/mern-auth",
+  secretOrKey: "secret",
+};
