@@ -1,24 +1,28 @@
-const JwtStrategy = require("passport-jwt").Strategy;
-const ExtractJwt = require("passport-jwt").ExtractJwt;
-const mongoose = require("mongoose");
-const User = mongoose.model("users");
-const keys = require("../config/keys");
+import { Strategy, ExtractJwt } from "passport-jwt";
+import { model } from "mongoose";
+import { key } from "./";
+const User = model("users");
 
 const opts = {};
-opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
-opts.secretOrKey = keys.secretOrKey;
 
-module.exports = passport => {
+opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
+opts.secretOrKey = key.secretOrKey;
+
+/**
+ *
+ * @param {import("passport").PassportStatic} passport
+ */
+export const passport = (passport) => {
   passport.use(
-    new JwtStrategy(opts, (jwt_payload, done) => {
+    new Strategy(opts, (jwt_payload, done) => {
       User.findById(jwt_payload.id)
-        .then(user => {
+        .then((user) => {
           if (user) {
             return done(null, user);
           }
           return done(null, false);
         })
-        .catch(err => console.log(err));
+        .catch((err) => console.log(err));
     })
   );
 };
