@@ -1,27 +1,24 @@
 import Validator from "validator";
-import isEmpty from "is-empty";
 
-module.exports = function validateLoginInput(data){
+module.exports = function validateLoginInput(data) {
+  let errors = {};
 
-    let errors = {};
+  data.email = data.email ?? "";
+  data.password = data.password ?? "";
 
-    data.email = !isEmpty(data.email) ? data.email : '';
-    data.password = !isEmpty(data.password) ? data.password : '';
+  //Email checks
+  if (Validator.isEmpty(data.email)) {
+    errors.email = "Email field is required";
+  } else if (!Validator.isEmail(data.email)) {
+    errors.email = "Email is invalid";
+  }
 
-    //Email checks
-    if(Validator.isEmpty(data.email)){
-        errors.email = "Email field is required";
-    } else if(!Validator.isEmail(data.email)){
-        errors.email = "Email is invalid";
-    }
+  //Password checks
+  if (Validator.isEmpty(data.password)) {
+    errors.password = "Password field is required";
+  }
 
-    //Password checks
-    if(Validator.isEmpty(data.password)){
-        errors.password = "Password field is required";
-    }
-
-    return{
-        errors,
-        isValid: isEmpty(errors)
-    };
+  return {
+    errors,
+  };
 };
